requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-serializer==1.3.4.4
urllib3==1.26.3
vk-api==11.9.1
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.2.0
python-dotenv

pillow~=10.2.0